fix(snack-bar): SimpleSnackBar not being exported
* Fixes the `SimpleSnackBar` class not being exported.
* Moves the `MdSnackBarModule` to the `index.ts` for consistency with other modules.

Fixes #3010.
