+++ src/lib/module.ts
@@ -27,7 +27,7 @@ import {MdIconModule} from './icon/index';
 import {MdProgressSpinnerModule} from './progress-spinner/index';
 import {MdProgressBarModule} from './progress-bar/index';
 import {MdInputModule} from './input/index';
-import {MdSnackBarModule} from './snack-bar/snack-bar';
+import {MdSnackBarModule} from './snack-bar/index';
 import {MdTabsModule} from './tabs/index';
 import {MdToolbarModule} from './toolbar/index';
 import {MdTooltipModule} from './tooltip/index';
