+++ src/lib/snack-bar/snack-bar.ts
@@ -1,24 +1,12 @@
-import {
-  NgModule,
-  ModuleWithProviders,
-  Injectable,
-  ComponentRef,
-  Optional,
-  SkipSelf,
-} from '@angular/core';
+import {Injectable, ComponentRef, Optional, SkipSelf} from '@angular/core';
 import {
   ComponentType,
   ComponentPortal,
   Overlay,
-  OverlayModule,
   OverlayRef,
   OverlayState,
-  PortalModule,
   LiveAnnouncer,
-  CompatibilityModule,
-  LIVE_ANNOUNCER_PROVIDER,
 } from '../core';
-import {CommonModule} from '@angular/common';
 import {MdSnackBarConfig} from './snack-bar-config';
 import {MdSnackBarRef} from './snack-bar-ref';
 import {MdSnackBarContainer} from './snack-bar-container';
@@ -159,22 +147,4 @@ export class MdSnackBar {
  */
 function _applyConfigDefaults(config: MdSnackBarConfig): MdSnackBarConfig {
   return extendObject(new MdSnackBarConfig(), config);
-}
-
-
-@NgModule({
-  imports: [OverlayModule, PortalModule, CommonModule, CompatibilityModule],
-  exports: [MdSnackBarContainer, CompatibilityModule],
-  declarations: [MdSnackBarContainer, SimpleSnackBar],
-  entryComponents: [MdSnackBarContainer, SimpleSnackBar],
-  providers: [MdSnackBar, LIVE_ANNOUNCER_PROVIDER]
-})
-export class MdSnackBarModule {
-  /** @deprecated */
-  static forRoot(): ModuleWithProviders {
-    return {
-      ngModule: MdSnackBarModule,
-      providers: []
-    };
-  }
 }
